Adding parallel dataStorageLayer classes
Adding a set of mirroring classes in the
rifDataLoaderTool.dataStorageLayer.ms package that will do what the
corresponding ones in the pg package will do.  Initially they will be
identical, but further analysis should mean some of them are altered to
consider specific syntax or function needs of SQL Server.