Adding configuration class
Adding a new configuration class that will eventually replace the need
for DataLoaderToolSettings.  It is meant to reflect the changes we're
putting in where configuration elements are grouped together and ordered
based on the categories of database settings, geography meta data,
health themes, rif data types, denominators, numerators, and covariates.
Whereas before many of the elements could be entered in one order or
another, now we're trying to set an ordering.