Adding convenience method
Most of the time when we are using the query formatter classes, the
parameter values are represented by question mark '?' characters, in
order to make them appropriate for setting PreparedStatement objects in
Java.  But when we use the same query formatters to generate load
scripts, there is a need to insert literal values (at least there is a
need in the Insert Query formatter).