Supporting re-run of load scripts
Modified the scripts that the data loader tool generates for PostgreSQL
so that it can be run and re-run. This requires a preliminary bit of
code that can delete tables and rows that were affected by the previous
run of the script.  In future, we'll have to give more thought about
whether tables should be overwritten or not.